+++ gpu/include/kernels/model.cuh
@@ -8,9 +8,9 @@
 
 constexpr u32 NO_TRAVEL = std::numeric_limits<u32>::max();
 constexpr u32 NO_LIMIT = std::numeric_limits<u32>::max();
-constexpr usize NO_PARENT = std::numeric_limits<usize>::max();
-constexpr usize NO_ROOM = std::numeric_limits<usize>::max();
-constexpr usize NO_CLASS_ASSIGNED = std::numeric_limits<usize>::max();
+constexpr u16 NO_PARENT = std::numeric_limits<u16>::max();
+constexpr u16 NO_ROOM = std::numeric_limits<u16>::max();
+constexpr u16 NO_CLASS_ASSIGNED = std::numeric_limits<u16>::max();
 
 namespace kernels {
 
@@ -37,32 +37,32 @@ struct RoomData {
 
 struct CourseData {
     thrust::device_vector<parser::CourseId> id;
-    thrust::device_vector<usize> configs_start;
-    thrust::device_vector<usize> configs_end;
+    thrust::device_vector<u16> configs_start;
+    thrust::device_vector<u16> configs_end;
 
     CourseData(const std::vector<parser::CourseId> &id,
-               const std::vector<usize> &configs_start,
-               const std::vector<usize> &configs_end);
+               const std::vector<u16> &configs_start,
+               const std::vector<u16> &configs_end);
 };
 
 struct ConfigData {
     thrust::device_vector<parser::ConfigId> id;
-    thrust::device_vector<usize> subparts_start;
-    thrust::device_vector<usize> subparts_end;
+    thrust::device_vector<u16> subparts_start;
+    thrust::device_vector<u16> subparts_end;
 
     ConfigData(const std::vector<parser::ConfigId> &id,
-               const std::vector<usize> &subparts_start,
-               const std::vector<usize> &subparts_end);
+               const std::vector<u16> &subparts_start,
+               const std::vector<u16> &subparts_end);
 };
 
 struct SubpartData {
     thrust::device_vector<parser::SubpartId> id;
-    thrust::device_vector<usize> classes_start;
-    thrust::device_vector<usize> classes_end;
+    thrust::device_vector<u16> classes_start;
+    thrust::device_vector<u16> classes_end;
 
     SubpartData(const std::vector<parser::SubpartId> &id,
-                const std::vector<usize> &classes_start,
-                const std::vector<usize> &classes_end);
+                const std::vector<u16> &classes_start,
+                const std::vector<u16> &classes_end);
 };
 
 struct ClassData {
@@ -70,24 +70,24 @@ struct ClassData {
     // U32_MAX if there's no limit
     thrust::device_vector<u32> limit;
     // USIZE_MAX if this class has no parent
-    thrust::device_vector<usize> parent;
+    thrust::device_vector<u16> parent;
     // indices into TimetableData::time_options
-    thrust::device_vector<usize> times_start;
-    thrust::device_vector<usize> times_end;
+    thrust::device_vector<u16> times_start;
+    thrust::device_vector<u16> times_end;
     // indices into TimetableData::room_options
     // rooms_start == rooms_end means the class doesn't need a room
-    thrust::device_vector<usize> rooms_start;
-    thrust::device_vector<usize> rooms_end;
+    thrust::device_vector<u16> rooms_start;
+    thrust::device_vector<u16> rooms_end;
     // indices into TimetableData::subparts
-    thrust::device_vector<usize> subpart_idx;
+    thrust::device_vector<u16> subpart_idx;
 
     ClassData(const std::vector<parser::ClassId> &id,
-              const std::vector<u32> &limit, const std::vector<usize> &parent,
-              const std::vector<usize> &times_start,
-              const std::vector<usize> &times_end,
-              const std::vector<usize> &rooms_start,
-              const std::vector<usize> &rooms_end,
-              const std::vector<usize> &subpart_idx);
+              const std::vector<u32> &limit, const std::vector<u16> &parent,
+              const std::vector<u16> &times_start,
+              const std::vector<u16> &times_end,
+              const std::vector<u16> &rooms_start,
+              const std::vector<u16> &rooms_end,
+              const std::vector<u16> &subpart_idx);
 };
 
 struct TimeOption {
@@ -99,23 +99,23 @@ struct TimeOption {
 };
 
 struct RoomOption {
-    thrust::device_vector<usize> room_idx;
+    thrust::device_vector<u16> room_idx;
     thrust::device_vector<u32> penalty;
 
-    RoomOption(const std::vector<usize> &room_idx,
+    RoomOption(const std::vector<u16> &room_idx,
                const std::vector<u32> &penalty);
 };
 
 struct StudentData {
     thrust::device_vector<parser::StudentId> id;
     // indices into TimetableData::courses
-    thrust::device_vector<usize> course_idxs;
+    thrust::device_vector<u16> course_idxs;
     // the courses wanted by student `i` begin at idx course_idxs_offsets[i]
     // and end at idx course_idxs_offsets[i + 1] - 1
     thrust::device_vector<usize> course_idxs_offsets;
 
     StudentData(const std::vector<parser::StudentId> &id,
-                const std::vector<usize> &course_idxs,
+                const std::vector<u16> &course_idxs,
                 const std::vector<usize> &course_idxs_offsets);
 };
 
@@ -123,14 +123,14 @@ struct DistributionData {
     // apparently CUDA supports std::variant so this should be fine
     thrust::device_vector<parser::DistributionKind> kind;
     // incides into TimetableData::classes
-    thrust::device_vector<usize> class_idxs;
+    thrust::device_vector<u16> class_idxs;
     // the idxs of classes taken into account by distribution `i`
     // begin at idx class_idxs_offsets[i] and end at idx class_idx_offsets[i + 1] - 1
     thrust::device_vector<usize> class_idxs_offsets;
     thrust::device_vector<Penalty> penalty;
 
     DistributionData(const std::vector<parser::DistributionKind> &kind,
-                     const std::vector<usize> &class_idxs,
+                     const std::vector<u16> &class_idxs,
                      const std::vector<usize> &class_idxs_offsets,
                      const std::vector<Penalty> &penalty);
 };
